Implemented a basic form  of Player collision checking, implemented bullet movement, and more.
Changed how the player data and bullet data are declared in the DS.
Converted GetPlayerInput to a procedure due to its length.
Added the ability to move the second player and shoot with both players.
Added very basic bounds checking for bullets so they don't continue moving forever.
The current implementation of player collision checking goes over every block of the player and checks for collisons. This however, is very costly and seems to cause major slowdowns whenever a bullet is active.
it checks player 1 and  then player 2 seperately.
The basic algorithm is as follows:
for each active bullet:
    for each block in the player
        If B.x = P.x && B.y =P.y
        // Do something (TBD)
        Stop looping