Impoved bullet-cactus collision checking to check for the cactus's xPos and xPos+1.
Since bullets move with  xVelocity = 2 by default, checking collisions with xPos alone would not work properly if the object's xPos is even and the bullet's start xPos is odd.
For P1 bullets for example, they would always spawn and x=3, then move with xVelocity = 2. This would cause them to move with xPos always odd, if the cactus's xPos was even, the xPos check would always fail and no colission would be detected.
Now, for cacti at least, we check both xPos and xPos + 1, this should check both even and odd xPos for the object/bullet and make the bullet's colission checking independent of it's starting xPos.